pyproject.toml
[tool.poetry]
name = "pyramid_cloudflare_access"
version = "1.0.0"
description = "A bunch of helpers for successfully running Pyramid on Heroku."
readme = "README.rst"
include = [ "CHANGES.rst", "COPYING.txt" ]
classifiers = [
    "Operating System :: OS Independent",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Framework :: Pylons",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: WSGI",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [
    { include = "pyramid_cloudflare_access" },
]
authors = [ "Niteo <info@niteo.co>" ]
license = "BSD-3-Clause"
homepage = "https://github.com/teamniteo/pyramid_cloudflare_access"
keywords = [ "pyramid", "cloudflare", "pylons", "web" ]

[tool.poetry.dependencies]
python = ">=3.6"
pyramid = ">=1.7"
pyjwt = "*"

[tool.poetry.dev-dependencies]
typecov = "*"
structlog = "*"
responses = "*"
pytest-mock = "*"
pytest-cov = "*"
pytest = "*"
mypy = "*"
mock = "*"
isort = "*"
docutils = "*"
black = "*"
autoflake = "*"
"zope.testing" = "*"

